@@ -3,7 +3,7 @@ package gr.academic.city.mc.studentmanagement;
 import android.content.Intent;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.RadioGroup;
 
